+++ src/bun.js/bindings/generated_classes_list.zig
@@ -4,11 +4,11 @@ pub const Classes = struct {
     pub const Blob = JSC.WebCore.Blob;
     pub const CryptoHasher = JSC.API.Bun.Crypto.CryptoHasher;
     pub const Dirent = JSC.Node.Dirent;
-    pub const Expect = JSC.Jest.Expect;
-    pub const ExpectAny = JSC.Jest.ExpectAny;
-    pub const ExpectAnything = JSC.Jest.ExpectAnything;
-    pub const ExpectStringContaining = JSC.Jest.ExpectStringContaining;
-    pub const ExpectStringMatching = JSC.Jest.ExpectStringMatching;
+    pub const Expect = JSC.Expect.Expect;
+    pub const ExpectAny = JSC.Expect.ExpectAny;
+    pub const ExpectAnything = JSC.Expect.ExpectAnything;
+    pub const ExpectStringContaining = JSC.Expect.ExpectStringContaining;
+    pub const ExpectStringMatching = JSC.Expect.ExpectStringMatching;
     pub const FileSystemRouter = JSC.API.FileSystemRouter;
     pub const Bundler = JSC.API.JSBundler;
     pub const JSBundler = Bundler;
