+++ src/bun.js/javascript.zig
@@ -653,6 +653,10 @@ pub const VirtualMachine = struct {
         this.eventLoop().waitForPromise(promise);
     }
 
+    pub fn waitForPromiseWithTimeout(this: *VirtualMachine, promise: JSC.AnyPromise, timeout: u32) bool {
+        return this.eventLoop().waitForPromiseWithTimeout(promise, timeout);
+    }
+
     pub fn waitForTasks(this: *VirtualMachine) void {
         this.eventLoop().waitForTasks();
     }
@@ -957,6 +961,7 @@ pub const VirtualMachine = struct {
     }
 
     pub fn refCountedStringWithWasNew(this: *VirtualMachine, new: *bool, input_: []const u8, hash_: ?u32, comptime dupe: bool) *JSC.RefString {
+        JSC.markBinding(@src());
         const hash = hash_ orelse JSC.RefString.computeHash(input_);
 
         var entry = this.ref_strings.getOrPut(hash) catch unreachable;
