+++ src/cli/test_command.zig
@@ -41,7 +41,7 @@ const HTTPThread = @import("root").bun.HTTP.HTTPThread;
 const JSC = @import("root").bun.JSC;
 const jest = JSC.Jest;
 const TestRunner = JSC.Jest.TestRunner;
-const Snapshots = JSC.Jest.Snapshots;
+const Snapshots = JSC.Snapshot.Snapshots;
 const Test = TestRunner.Test;
 const NetworkThread = @import("root").bun.HTTP.NetworkThread;
 const uws = @import("root").bun.uws;
